+++ vclib-aries/src/commonTest/kotlin/at/asitplus/wallet/lib/aries/IssueCredentialMessengerConcurrentTest.kt
@@ -1,16 +1,16 @@
 package at.asitplus.wallet.lib.aries
 
+import at.asitplus.wallet.lib.agent.Holder
 import at.asitplus.wallet.lib.agent.HolderAgent
 import at.asitplus.wallet.lib.agent.Issuer
 import at.asitplus.wallet.lib.agent.IssuerAgent
 import at.asitplus.wallet.lib.agent.KeyPairAdapter
 import at.asitplus.wallet.lib.agent.RandomKeyPairAdapter
+import at.asitplus.wallet.lib.agent.SubjectCredentialStore
 import at.asitplus.wallet.lib.data.AtomicAttribute2023
 import at.asitplus.wallet.lib.data.ConstantIndex
 import com.benasher44.uuid.uuid4
 import io.kotest.core.spec.style.FreeSpec
-import io.kotest.matchers.collections.shouldBeEmpty
-import io.kotest.matchers.collections.shouldNotBeEmpty
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
 import kotlinx.coroutines.coroutineScope
@@ -83,12 +83,12 @@ class IssueCredentialMessengerConcurrentTest : FreeSpec() {
         return issuedCredential
     }
 
-    private fun assertAtomicVc(issuedCredentials: IssueCredentialProtocolResult) {
-        issuedCredentials.acceptedVcJwt.shouldNotBeEmpty()
-        issuedCredentials.acceptedVcJwt.map { it.vc.credentialSubject }.forEach {
-            it.shouldBeInstanceOf<AtomicAttribute2023>()
-        }
-        issuedCredentials.rejected.shouldBeEmpty()
+    private fun assertAtomicVc(issuedCredential: IssueCredentialProtocolResult) {
+        val credential = issuedCredential.getOrThrow()
+        credential.shouldBeInstanceOf<Holder.StoredCredential.Vc>()
+        val storeEntry = credential.storeEntry
+        storeEntry.shouldBeInstanceOf<SubjectCredentialStore.StoreEntry.Vc>()
+        storeEntry.vc.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
     }
 
 }